samd: Mount (Q)SPI flash at boot time as file system.
Increasing the size of the file system. For SAMD51 devices it increases
as well the speed of file access and the code execution substantially.

For SAMD21 devices, the Board flash signals must be named in pins.csv
as FLASH_MOSI, FLASH_MISO, FLASH_SCK, FLASH_CS for creating the SPI
object in _boot.py. The SPI number must be set be the define for
MICROPY_HW_SPIFLASH_ID in mpconfigboard.h.

Even if using the external flash would be possible with the
Sparkfun SAMD51 Thing, there is almost no benefit doing so. The external
flash is smaller than the internal one, and there is no MCU speed benefit.

At boards with large flash, VfsLfs2 works slow with sometimes long periods
of flash reorganisation. For these, VfsLfs1 or  VfsFat is faster.

Rename the QSPI pins to QSPI_xxxx instead of FLASH_xxx.