samd: Rearrange the mcu specific loader files.
Such that they are easier to adapt. The maximum code size is set by:

MICROPY_HW_CODESIZE=xxxK

in mpconfigmcu.mk for the MCU family as default or
in mpconfigboard.mk for a specific board.

Setting the maximum code size allows the loader to raise and error
if the code gets larger than the space dedicated for it.