shared/runtime/softtimer.c: Use consistently the same clock source.
Before, both uwTick and mp_hal_ticks_ms() were used as clock source.
That assumes, that these two are synchronous and start with the same
value, which may be not the case for all ports. If the lag between
uwTick and mp_hal_ticks_ms() is larger than the timer interval, the
timer would either rush up until the times as synchronous, or not
start until uwTick wraps over.

Using consistently uwTick removes the potential discrepancy.